apply feeds
