+++ src/modules/Channel/components/ChannelUI/index.tsx
@@ -9,9 +9,6 @@ import { GroupChannelUIView } from '../../../GroupChannel/components/GroupChanne
 import ChannelHeader from '../ChannelHeader';
 import MessageList from '../MessageList';
 import MessageInputWrapper from '../MessageInputWrapper';
-import { getSuggestedReplies } from '../../../../utils';
-import type { UserMessage } from '@sendbird/chat/message';
-import useSendbirdStateContext from '../../../../hooks/useSendbirdStateContext';
 
 export interface ChannelUIProps {
   isLoading?: boolean;
@@ -31,27 +28,12 @@ export interface ChannelUIProps {
 }
 
 const ChannelUI = (props: ChannelUIProps) => {
-  const { config } = useSendbirdStateContext();
   const context = useChannelContext();
   const {
     channelUrl,
     isInvalid,
     loading,
-    localMessages,
-    currentGroupChannel,
   } = context;
-
-  const lastMessage = currentGroupChannel?.lastMessage;
-  const isLastMessageSuggestedRepliesEnabled = config?.groupChannel?.enableSuggestedReplies
-    && lastMessage
-    && getSuggestedReplies(lastMessage).length > 0
-    && (
-      !localMessages
-      || localMessages.length === 0
-      || localMessages.every((message) => (message as UserMessage).sendingStatus === 'succeeded')
-    );
-  const disableMessageInput = isLastMessageSuggestedRepliesEnabled
-    && !!lastMessage.extendedMessagePayload?.['disable_chat_input'];
 
   return (
     <GroupChannelUIView
@@ -64,7 +46,7 @@ const ChannelUI = (props: ChannelUIProps) => {
       renderMessageList={(props) => (<MessageList {...props} />)}
       renderMessageInput={() => (
         props.renderMessageInput?.()
-        ?? <MessageInputWrapper {...props} disabled={disableMessageInput} />
+        ?? <MessageInputWrapper {...props} />
       )}
     />
   );
