+++ src/modules/GroupChannel/components/GroupChannelUI/index.tsx
@@ -8,9 +8,6 @@ import { useGroupChannelContext } from '../../context/GroupChannelProvider';
 import { GroupChannelUIView } from './GroupChannelUIView';
 import type { MessageContentProps } from '../../../../ui/MessageContent';
 import MessageInputWrapper from '../MessageInputWrapper';
-import { useIIFE } from '@sendbird/uikit-tools';
-import { getSuggestedReplies, isSendableMessage } from '../../../../utils';
-import useSendbirdStateContext from '../../../../hooks/useSendbirdStateContext';
 
 export interface GroupChannelUIProps {
   renderPlaceholderLoader?: () => React.ReactElement;
@@ -30,25 +27,12 @@ export interface GroupChannelUIProps {
 }
 
 export const GroupChannelUI = (props: GroupChannelUIProps) => {
-  const { config } = useSendbirdStateContext();
   const context = useGroupChannelContext();
   const {
     currentChannel,
     channelUrl,
     loading,
-    messages,
   } = context;
-
-  const lastMessage = messages[messages.length - 1];
-  const isLastMessageSuggestedRepliesEnabled = useIIFE(() => {
-    if (!config?.groupChannel?.enableSuggestedReplies) return false;
-    if (getSuggestedReplies(lastMessage).length === 0) return false;
-    if (lastMessage && isSendableMessage(lastMessage) && lastMessage.sendingStatus !== 'succeeded') return false;
-
-    return true;
-  });
-  const disableMessageInput = isLastMessageSuggestedRepliesEnabled
-    && !!lastMessage.extendedMessagePayload?.['disable_chat_input'];
 
   return (
     <GroupChannelUIView
@@ -59,7 +43,7 @@ export const GroupChannelUI = (props: GroupChannelUIProps) => {
       isInvalid={channelUrl && !currentChannel}
       renderMessageInput={() => (
         props.renderMessageInput?.()
-        ?? <MessageInputWrapper {...props} disabled={disableMessageInput} />
+        ?? <MessageInputWrapper {...props} />
       )}
     />
   );
